mmp2p add lazy loading for sidecard
